requirements.txt
android-backup==0.2.0
appdirs==1.4.4
attrs==21.2.0
cffi==1.14.5
click==7.1.2
construct==2.10.67
croniter==0.3.37
cryptography==3.4.7
defusedxml==0.6.0
ifaddr==0.1.7
natsort==7.1.1
netifaces==0.10.9
pycparser==2.20
python-dateutil==2.8.1
python-miio==0.5.6
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
tqdm==4.60.0
websockets==9.0.1
zeroconf==0.30.0